+++ cmd/app/commands/create_kek_test.go
@@ -7,18 +7,18 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	cryptoDomain "github.com/allisson/secrets/internal/crypto/domain"
-	cryptoMocks "github.com/allisson/secrets/internal/crypto/usecase/mocks"
+	"github.com/allisson/secrets/internal/keyring"
+	keyringMocks "github.com/allisson/secrets/internal/keyring/mocks"
 )
 
 func TestRunCreateKek(t *testing.T) {
 	ctx := context.Background()
 	logger := slog.Default()
-	masterKeyChain := cryptoDomain.NewMasterKeyChain("key1")
+	masterKeyChain := keyring.NewMasterKeyChain("key1")
 
 	t.Run("success", func(t *testing.T) {
-		mockUseCase := &cryptoMocks.MockKekUseCase{}
-		mockUseCase.On("Create", ctx, masterKeyChain, cryptoDomain.AESGCM).Return(nil)
+		mockUseCase := &keyringMocks.MockKekUseCase{}
+		mockUseCase.On("Create", ctx, masterKeyChain, keyring.AESGCM).Return(nil)
 
 		err := RunCreateKek(ctx, mockUseCase, masterKeyChain, logger, "aes-gcm")
 
@@ -27,7 +27,7 @@ func TestRunCreateKek(t *testing.T) {
 	})
 
 	t.Run("invalid-algorithm", func(t *testing.T) {
-		mockUseCase := &cryptoMocks.MockKekUseCase{}
+		mockUseCase := &keyringMocks.MockKekUseCase{}
 		err := RunCreateKek(ctx, mockUseCase, masterKeyChain, logger, "invalid")
 
 		require.Error(t, err)